+++ monty/__init__.py
@@ -5,9 +5,9 @@ Examples include useful utilities like transparent support for zipped files,
 useful design patterns such as singleton and cached_class, and many more.
 """
 
-__author__ = 'Shyue Ping Ong'
-__copyright__ = 'Copyright 2014, The Materials Virtual Lab'
-__version__ = '4.0.0'
-__maintainer__ = 'Shyue Ping Ong'
-__email__ = 'ongsp@ucsd.edu'
-__date__ = 'Sep 1 2020'
+__author__ = "Shyue Ping Ong"
+__copyright__ = "Copyright 2014, The Materials Virtual Lab"
+__version__ = "2021.8.17"
+__maintainer__ = "Shyue Ping Ong"
+__email__ = "ongsp@ucsd.edu"
+__date__ = "Oct 12 2020"