+++ monty/inspect.py
@@ -49,7 +49,7 @@ def caller_name(skip=2):
     stack = inspect.stack()
     start = 0 + skip
     if len(stack) < start + 1:
-        return ''
+        return ""
     parentframe = stack[start][0]
 
     name = []
@@ -59,13 +59,13 @@ def caller_name(skip=2):
     if module:
         name.append(module.__name__)
     # detect classname
-    if 'self' in parentframe.f_locals:
+    if "self" in parentframe.f_locals:
         # I don't know any way to detect call from the object method
         # XXX: there seems to be no way to detect static method call - it will
         #      be just a function call
-        name.append(parentframe.f_locals['self'].__class__.__name__)
+        name.append(parentframe.f_locals["self"].__class__.__name__)
     codename = parentframe.f_code.co_name
-    if codename != '<module>':  # top level usually
+    if codename != "<module>":  # top level usually
         name.append(codename)  # function or a method
     del parentframe
     return ".".join(name)