+++ monty/subprocess.py
@@ -11,12 +11,12 @@ from subprocess import Popen, PIPE
 from .string import is_string
 
 
-__author__ = 'Matteo Giantomass'
+__author__ = "Matteo Giantomass"
 __copyright__ = "Copyright 2014, The Materials Virtual Lab"
-__version__ = '0.1'
-__maintainer__ = 'Matteo Giantomassi'
-__email__ = 'gmatteo@gmail.com'
-__date__ = '10/26/14'
+__version__ = "0.1"
+__maintainer__ = "Matteo Giantomassi"
+__email__ = "gmatteo@gmail.com"
+__date__ = "10/26/14"
 
 
 class Command:
@@ -59,7 +59,7 @@ class Command:
         self.command = command
         self.process = None
         self.retcode = None
-        self.output, self.error = '', ''
+        self.output, self.error = "", ""
         self.killed = False
 
     def __str__(self):
@@ -76,20 +76,20 @@ class Command:
         def target(**kw):
             try:
                 # print('Thread started')
-                self.process = Popen(self.command, **kw)
-                self.output, self.error = self.process.communicate()
-                self.retcode = self.process.returncode
+                with Popen(self.command, **kw) as self.process:
+                    self.output, self.error = self.process.communicate()
+                    self.retcode = self.process.returncode
                 # print('Thread stopped')
             except Exception:
                 self.error = traceback.format_exc()
                 self.retcode = -1
 
         # default stdout and stderr
-        if 'stdout' not in kwargs:
-            kwargs['stdout'] = PIPE
+        if "stdout" not in kwargs:
+            kwargs["stdout"] = PIPE
 
-        if 'stderr' not in kwargs:
-            kwargs['stderr'] = PIPE
+        if "stderr" not in kwargs:
+            kwargs["stderr"] = PIPE
 
         # thread
         thread = threading.Thread(target=target, kwargs=kwargs)