+++ tests/test_dev.py
@@ -5,7 +5,6 @@ from monty.dev import deprecated, requires, get_ncpus, install_excepthook
 
 
 class A:
-
     @property
     def repl_prop(self):
         pass
@@ -17,9 +16,7 @@ class A:
 
 
 class DecoratorTest(unittest.TestCase):
-
     def test_deprecated(self):
-
         def func_a():
             pass
 
@@ -33,11 +30,10 @@ class DecoratorTest(unittest.TestCase):
             # Trigger a warning.
             func_b()
             # Verify some things
-            self.assertTrue(issubclass(w[0].category, DeprecationWarning))
-            self.assertIn('hello', str(w[0].message))
+            self.assertTrue(issubclass(w[0].category, FutureWarning))
+            self.assertIn("hello", str(w[0].message))
 
     def test_deprecated_property(self):
-
         class a(object):
             def __init__(self):
                 pass
@@ -49,31 +45,30 @@ class DecoratorTest(unittest.TestCase):
             @property  # type: ignore
             @deprecated(property_a)
             def property_b(self):
-                return 'b'
+                return "b"
 
             @deprecated(property_a)
             def func_a(self):
-                return 'a'
+                return "a"
 
         with warnings.catch_warnings(record=True) as w:
             # Cause all warnings to always be triggered.
             warnings.simplefilter("always")
             # Trigger a warning.
-            self.assertEqual(a().property_b, 'b')
+            self.assertEqual(a().property_b, "b")
             # Verify some things
-            self.assertTrue(issubclass(w[-1].category, DeprecationWarning))
+            self.assertTrue(issubclass(w[-1].category, FutureWarning))
 
         with warnings.catch_warnings(record=True) as w:
             # Cause all warnings to always be triggered.
             warnings.simplefilter("always")
             # Trigger a warning.
-            self.assertEqual(a().func_a(), 'a')
+            self.assertEqual(a().func_a(), "a")
             # Verify some things
-            self.assertTrue(issubclass(w[-1].category, DeprecationWarning))
+            self.assertTrue(issubclass(w[-1].category, FutureWarning))
 
     def test_deprecated_classmethod(self):
-
-        class a(object):
+        class A(object):
             def __init__(self):
                 pass
 
@@ -84,13 +79,34 @@ class DecoratorTest(unittest.TestCase):
             @classmethod
             @deprecated(classmethod_a)
             def classmethod_b(self):
-                return 'b'
+                return "b"
 
         with warnings.catch_warnings(record=True) as w:
             # Cause all warnings to always be triggered.
             warnings.simplefilter("always")
             # Trigger a warning.
-            self.assertEqual(a().classmethod_b(), 'b')
+            self.assertEqual(A().classmethod_b(), "b")
+            # Verify some things
+            self.assertTrue(issubclass(w[-1].category, FutureWarning))
+
+        class A(object):
+            def __init__(self):
+                pass
+
+            @classmethod
+            def classmethod_a(self):
+                pass
+
+            @classmethod
+            @deprecated(classmethod_a, category=DeprecationWarning)
+            def classmethod_b(self):
+                return "b"
+
+        with warnings.catch_warnings(record=True) as w:
+            # Cause all warnings to always be triggered.
+            warnings.simplefilter("always")
+            # Trigger a warning.
+            self.assertEqual(A().classmethod_b(), "b")
             # Verify some things
             self.assertTrue(issubclass(w[-1].category, DeprecationWarning))