+++ tests/test_os.py
@@ -5,11 +5,10 @@ import platform
 from monty.os.path import which, zpath, find_exts
 from monty.os import cd, makedirs_p
 
-test_dir = os.path.join(os.path.dirname(__file__), 'test_files')
+test_dir = os.path.join(os.path.dirname(__file__), "test_files")
 
 
 class PathTest(unittest.TestCase):
-
     @unittest.skipIf(platform.system() == "Windows", "Skip on windows")
     def test_which(self):
         try:
@@ -28,14 +27,17 @@ class PathTest(unittest.TestCase):
     def test_find_exts(self):
         self.assertTrue(len(find_exts(os.path.dirname(__file__), "py")) >= 18)
         self.assertEqual(len(find_exts(os.path.dirname(__file__), "bz2")), 2)
-        self.assertEqual(len(find_exts(os.path.dirname(__file__), "bz2",
-                                       exclude_dirs="test_files")), 0)
-        self.assertEqual(len(find_exts(os.path.dirname(__file__), "bz2",
-                                       include_dirs="test_files")), 2)
+        self.assertEqual(
+            len(find_exts(os.path.dirname(__file__), "bz2", exclude_dirs="test_files")),
+            0,
+        )
+        self.assertEqual(
+            len(find_exts(os.path.dirname(__file__), "bz2", include_dirs="test_files")),
+            2,
+        )
 
 
 class CdTest(unittest.TestCase):
-
     def test_cd(self):
         with cd(test_dir):
             self.assertTrue(os.path.exists("empty_file.txt"))
@@ -48,7 +50,6 @@ class CdTest(unittest.TestCase):
 
 
 class Makedirs_pTest(unittest.TestCase):
-
     def setUp(self):
         self.test_dir_path = os.path.join(test_dir, "test_dir")