+++ tests/test_tempfile.py
@@ -5,12 +5,10 @@ from io import open
 
 from monty.tempfile import ScratchDir
 
-test_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)),
-                        'test_files')
+test_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "test_files")
 
 
 class ScratchDirTest(unittest.TestCase):
-
     def setUp(self):
         self.cwd = os.getcwd()
         os.chdir(test_dir)
@@ -20,12 +18,15 @@ class ScratchDirTest(unittest.TestCase):
     def test_with_copy(self):
         # We write a pre-scratch file.
         with open("pre_scratch_text", "w") as f:
-            f.write(u"write")
+            f.write("write")
 
-        with ScratchDir(self.scratch_root, copy_from_current_on_enter=True,
-                        copy_to_current_on_exit=True) as d:
+        with ScratchDir(
+            self.scratch_root,
+            copy_from_current_on_enter=True,
+            copy_to_current_on_exit=True,
+        ) as d:
             with open("scratch_text", "w") as f:
-                f.write(u"write")
+                f.write("write")
             files = os.listdir(d)
             self.assertIn("scratch_text", files)
             self.assertIn("empty_file.txt", files)
@@ -46,10 +47,13 @@ class ScratchDirTest(unittest.TestCase):
 
     def test_no_copy(self):
 
-        with ScratchDir(self.scratch_root, copy_from_current_on_enter=False,
-                        copy_to_current_on_exit=False) as d:
+        with ScratchDir(
+            self.scratch_root,
+            copy_from_current_on_enter=False,
+            copy_to_current_on_exit=False,
+        ) as d:
             with open("scratch_text", "w") as f:
-                f.write(u"write")
+                f.write("write")
             files = os.listdir(d)
             self.assertIn("scratch_text", files)
             self.assertNotIn("empty_file.txt", files)